+++ src/zepben/eas/client/work_package.py
@@ -1,0 +1,16 @@
+#  Copyright 2020 Zeppelin Bend Pty Ltd
+#
+#  This Source Code Form is subject to the terms of the Mozilla Public
+#  License, v. 2.0. If a copy of the MPL was not distributed with this
+#  file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+from dataclasses import dataclass
+from typing import List
+
+__all__ = ["WorkPackageConfig"]
+
+
+@dataclass
+class WorkPackageConfig:
+    """ A data class representing the configuration for a hosting capacity work package """
+    feeders: List[str]