+++ changelog.md
@@ -1,16 +1,21 @@
 ## [0.7.0]
 
 ##### Breaking Changes
+
 * None.
 
 ##### New Features
-* None.
+
+* Added basic client method to run hosting capacity work packages
 
 ##### Enhancements
+
 * None.
 
 ##### Fixes
+
 * None.
 
 ##### Notes
+
 * None.